+++ dart/geocore/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [Apply very_good_analysis 2.4.0+ lint rules #36](https://github.com/navibyte/geospatial/issues/36)
 - [Enhance methods of Point returning coordinate values as String #37](https://github.com/navibyte/geospatial/issues/37)
+- [Add fromText and toText serialization to Point classes #38](https://github.com/navibyte/geospatial/issues/38)
 
 ## 0.7.1